# application/api-spec/src/main/resources/openapi/hotel-reservation.yaml
openapi: 3.0.3
info:
  title: Hotel Reservation API
  version: 1.0.0
  description: |
    Hotel Reservation backend API for room availability search, reservation management,
    and staff stay operations.
servers:
  - url: http://localhost:8080
paths:
  /hotels:
    get:
      tags:
        - Hotels
      operationId: listHotels
      summary: List active hotels
      security:
        - bearerAuth: []
      parameters:
        - name: city
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Active hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /hotels/{hotelId}:
    get:
      tags:
        - Hotels
      operationId: getHotelDetails
      summary: Get active hotel details
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Hotel details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HotelResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Hotel not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /hotels/{hotelId}/services:
    get:
      tags:
        - Hotels
      operationId: listHotelServices
      summary: List active hotel services
      security:
        - bearerAuth: []
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Active hotel services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HotelServiceOfferingResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Hotel not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /hotels/{hotelId}/room-types/{roomTypeId}/availability-calendar:
    get:
      tags:
        - Rooms
      operationId: getRoomTypeAvailabilityCalendar
      summary: Get room type availability calendar
      description: Returns one item for every date in the inclusive [from, to] range. Each item describes whether at least one physical room of the requested room type is available for that night. Check-out dates are not counted as occupied nights. The date range cannot exceed 90 days.
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: roomTypeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Room type availability calendar
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomTypeAvailabilityCalendarDayResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Hotel or room type not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reservations:
    get:
      tags:
        - Reservations
      operationId: listReservations
      summary: List reservations
      description: Returns all reservations for staff and administrators.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 100
      responses:
        '200':
          description: Reservation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Reservations
      operationId: createReservation
      summary: Create authenticated guest reservation
      description: Creates a reservation for the guest profile linked to the authenticated GUEST JWT. The request body must not contain guestId.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
            examples:
              familyWithPet:
                value:
                  hotelId: 1001
                  roomTypeId: 1102
                  checkIn: '2026-05-10'
                  checkOut: '2026-05-12'
                  adults: 2
                  childrenAges: [7, 1]
                  pets:
                    - type: DOG
                      size: SMALL
                      weightKg: 8
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReservationResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /public/reservations:
    post:
      tags:
        - Reservations
      operationId: createPublicReservation
      summary: Create anonymous guest checkout reservation
      description: Creates a reservation without JWT. Backend creates or finds the Guest profile from contact data; guestId must not be supplied by the caller.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicCreateReservationRequest'
            examples:
              anonymousCheckout:
                value:
                  hotelId: 1001
                  roomTypeId: 1102
                  firstName: Anna
                  lastName: Tester
                  email: anna.tester@example.com
                  phone: '+421900000011'
                  checkIn: '2026-05-10'
                  checkOut: '2026-05-12'
                  adults: 2
                  childrenAges: [7]
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateReservationResponse'
        '400':
          description: Validation error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Authenticated callers must use POST /reservations
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservations/{reservationId}:
    get:
      tags:
        - Reservations
      operationId: getReservation
      summary: Get reservation by id
      security:
        - bearerAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reservation found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reservation not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservations/{reservationId}/cancel:
    post:
      tags:
        - Reservations
      operationId: cancelReservation
      summary: Cancel reservation by id
      security:
        - bearerAuth: []
      parameters:
        - name: reservationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Reservation cancelled
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reservation not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HotelResponse:
      type: object
      required:
        - hotelId
        - name
        - city
        - country
        - address
        - stars
        - status
        - childrenAllowed
        - petsAllowed
        - infantMaxAge
        - childMaxAge
        - adultEquivalentAge
      properties:
        hotelId:
          type: integer
          format: int64
        name:
          type: string
        city:
          type: string
        country:
          type: string
        address:
          type: string
        stars:
          type: integer
          minimum: 1
          maximum: 5
        description:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE, UNDER_MAINTENANCE]
        childrenAllowed:
          type: boolean
          description: Whether this hotel accepts children and infants. If false, childrenAges must be empty when searching or creating a reservation.
        petsAllowed:
          type: boolean
        infantMaxAge:
          type: integer
          description: Inclusive maximum age classified as an infant for this hotel. Infants are submitted inside childrenAges.
        childMaxAge:
          type: integer
          description: Inclusive maximum age that counts as a child for this hotel. Older minors from adultEquivalentAge count against adult occupancy instead.
        adultEquivalentAge:
          type: integer
          description: Inclusive age from which a minor submitted in childrenAges counts as an adult for occupancy limits. Must be less than 18.
    HotelServiceOfferingResponse:
      type: object
      required:
        - serviceOfferingId
        - hotelId
        - code
        - name
        - priceAmount
        - priceCurrency
        - active
      properties:
        serviceOfferingId:
          type: integer
          format: int64
        hotelId:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        description:
          type: string
        priceAmount:
          type: number
        priceCurrency:
          type: string
        active:
          type: boolean
        availabilityRule:
          type: string
...
    BookingPet:
      type: object
      required:
        - type
        - size
        - weightKg
      properties:
        type:
          type: string
          enum: [DOG, CAT, OTHER]
        size:
          type: string
          enum: [SMALL, MEDIUM, LARGE]
        weightKg:
          type: number
          minimum: 0
...
    CreateReservationRequest:
      type: object
      required:
        - hotelId
        - roomTypeId
        - checkIn
        - checkOut
        - adults
      properties:
        hotelId:
          type: integer
          format: int64
        roomTypeId:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        adults:
          type: integer
          minimum: 1
        childrenAges:
          type: array
          description: Ages of all accompanying minors, including infants. The backend classifies each age using the selected hotel's infantMaxAge, childMaxAge, and adultEquivalentAge; infants count against maxInfants, children count against maxChildren, and adult-equivalent minors count against maxAdults/maxTotalGuests.
          items:
            type: integer
            minimum: 0
        pets:
          type: array
          items:
            $ref: '#/components/schemas/BookingPet'
        contactEmail:
          type: string
          format: email
          nullable: true
          description: Per-booking contact email. When null, backend uses the Guest profile email.
        contactPhone:
          type: string
          nullable: true
          maxLength: 32
          description: Per-booking contact phone. When null, backend falls back to the Guest profile phone when available.
        specialRequests:
          type: string
          nullable: true
          maxLength: 500
          description: Free-form note from the guest to the hotel. Stored with the reservation; not a binding promise.
        serviceOfferings:
          type: array
          items:
            $ref: '#/components/schemas/ReservationServiceSelection'
    PublicCreateReservationRequest:
      type: object
      required:
        - hotelId
        - roomTypeId
        - firstName
        - lastName
        - email
        - phone
        - checkIn
        - checkOut
        - adults
      properties:
        hotelId:
          type: integer
          format: int64
        roomTypeId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        adults:
          type: integer
          minimum: 1
        childrenAges:
          type: array
          description: Ages of all accompanying minors, including infants. The backend classifies each age using the selected hotel's infantMaxAge, childMaxAge, and adultEquivalentAge; infants count against maxInfants, children count against maxChildren, and adult-equivalent minors count against maxAdults/maxTotalGuests.
          items:
            type: integer
            minimum: 0
        pets:
          type: array
          items:
            $ref: '#/components/schemas/BookingPet'
        serviceOfferings:
          type: array
          items:
            $ref: '#/components/schemas/ReservationServiceSelection'
...
    ReservationServiceSelection:
      type: object
      required:
        - serviceOfferingId
        - quantity
      properties:
        serviceOfferingId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    CreateReservationResponse:
      allOf:
        - $ref: '#/components/schemas/ReservationResponse'
    ReservationResponse:
      type: object
      required:
        - reservationId
        - hotelId
        - guestId
        - roomTypeId
        - checkIn
        - checkOut
        - adults
        - childrenAges
        - pets
        - status
        - createdAt
        - createdBy
      properties:
        reservationId:
          type: string
        hotelId:
          type: integer
          format: int64
        guestId:
          type: integer
          format: int64
        roomId:
          type: integer
          format: int64
          nullable: true
        roomTypeId:
          type: integer
          format: int64
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        adults:
          type: integer
        childrenAges:
          type: array
          description: Ages of all accompanying minors exactly as submitted, including infants. Infant/child classification is derived from the hotel's age policy.
          items:
            type: integer
        pets:
          type: array
          items:
            $ref: '#/components/schemas/BookingPet'
        contactEmail:
          type: string
          format: email
          nullable: true
        contactPhone:
          type: string
          nullable: true
          maxLength: 32
        specialRequests:
          type: string
          nullable: true
          maxLength: 500
        basePriceAmount:
          type: number
        basePriceCurrency:
          type: string
        servicesPriceAmount:
          type: number
        servicesPriceCurrency:
          type: string
        discountAmount:
          type: number
        discountCurrency:
          type: string
        finalPriceAmount:
          type: number
        finalPriceCurrency:
          type: string
        serviceItems:
          type: array
          items:
            $ref: '#/components/schemas/ReservationServiceItemResponse'
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        createdBy:
          type: string
    ReservationServiceItemResponse:
      type: object
      required:
        - serviceOfferingId
        - serviceName
        - priceAmount
        - priceCurrency
        - quantity
        - totalPriceAmount
        - totalPriceCurrency
      properties:
        serviceOfferingId:
          type: integer
          format: int64
        serviceName:
          type: string
        priceAmount:
          type: number
        priceCurrency:
          type: string
        quantity:
          type: integer
        totalPriceAmount:
          type: number
        totalPriceCurrency:
          type: string
...
    RoomTypeAvailabilityCalendarDayResponse:
      type: object
      required:
        - date
        - availableCount
        - available
      properties:
        date:
          type: string
          format: date
        availableCount:
          type: integer
        available:
          type: boolean
    ErrorResponse:
      type: object
      description: RFC 7807 problem details response returned by the REST layer.
      properties:
        type:
          type: string
          example: about:blank
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          format: int32
          example: 400
        detail:
          type: string
          example: checkOut must be after checkIn
        instance:
          type: string
          example: /reservations